Update rfc-002-working-in-the-open.md
Note 1: I would challenge this as DB and field information can be of use to attackers (mainly on web facing apps), often not directly, but the enumeration of this type of data is often part of the deeper penetration of the attack. Maybe an extra comment on the lines of 'although consideration must be taken on the value to a potential attacker as to whether appropriate to share in the open' 

Apert from this I cannot see other issues and like what you have created so far. I will do further thinking about whether there are additional info to add.